requirements.txt
streamlit
streamlit-image-coordinates
Pillow
numpy